Move Body component as molecule
